+++ src/modules/partners/services/DeletePartnersService.ts
@@ -1,3 +1,4 @@
+import AdsRepository from '@modules/ads/typeorm/repositories/AdsRepository';
 import AppError from '@shared/errors/AppError';
 import { getCustomRepository } from 'typeorm';
 import PartnersRepositories from '../typeorm/repositories/PartnersRepositories';
@@ -9,9 +10,15 @@ interface IRequest {
 class DeletePartnersService {
   public async execute({ id }: IRequest): Promise<void> {
     const partnerssRepository = getCustomRepository(PartnersRepositories);
+    const adsRepository = getCustomRepository(AdsRepository);
     const partners = await partnerssRepository.findOne(id);
     if (!partners) {
       throw new AppError('Parceiro não encontrado.');
+    }
+
+    const adsExists = await adsRepository.findById(partners.id);
+    if (adsExists?.id_partner) {
+      throw new AppError('Existe anuncio');
     }
 
     await partnerssRepository.remove(partners);
